Added capability to return the first matching rule during bulk evaluation (#24)
Currently HopeLangEngine support evaluating rules
* one-by-one
* bulk evaluation which will evaluate all rules irrespective of how many
of them match

Many a times, a client would want to evaluate a bunch of rules and just
wants to know if any of the rule matches instead of all the matching
rules. This PR aims to add another public method for evaluation
`evaluateFirst` to achieve the same. If any rule matches, that rule's
index in the original list would returned otherwise an
`Optional.empty()` would be returned

Signed-off-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>